mgw: Split CRCX read-only validations into its own function
The SDP parsing is moved to a step beforehand, so that validation of
full message can be done in a subsequent step. That validation step is
moved into a function to give some air to handle_create_con() and easily
spot the logic.

The logic based on force_realloc is split so that code modifying the
object is moved to the later update step.

Change-Id: I639ad0a25a0af4a4637045ca8bf61e436a789426